apis/tiktok-api/src/requirements.txt
# Web Framework
flask==3.0.3
gunicorn==22.0.0

# Data Processing
pandas==2.2.3

# Logging
loguru==0.7.2

# HTTP Requests
requests==2.32.3

# Google Cloud
google-cloud-bigquery==3.25.0
google-auth==2.34.0
google-api-core==2.19.2

# BigQuery dependencies
pyarrow==17.0.0
db-dtypes==1.3.0

# Orchestration
apache-airflow>=2.8.0
apache-airflow-providers-google
apache-airflow-providers-http
